+++ contrib/plom-make-scrap-sheets.py
@@ -4,12 +4,13 @@
 # Copyright (C) 2024 Philip D Loewen
 # Copyright (C) 2024 Andrew Rechnitzer
 # Copyright (C) 2025 Deep Shah
+# Copyright (C) 2025 Colin B. Macdonald
 
 """Generate a PDF with title text and QR codes both indicating scrap.
 
 All parameters are optional and have sensible defaults. Saying simply
     python3 plom-make-scrap-sheets.py
-will create (or overwrite!) a 2-page file named PLOM-scrap.pdf suitable
+will create (or overwrite!) a 2-page file named `Plom-scrap.pdf` suitable
 for double-sided printing and mass duplication.
 
 Simple mode:
@@ -37,8 +38,6 @@ You can even insert a Python integer-formatting code in those braces.
 E.g., you can say
     python3 plom-make-scrap-sheets.py -t "Scrap paper (seq {:04d})" -n 5
 """
-
-from __future__ import annotations
 
 import argparse
 
@@ -155,8 +154,8 @@ def configure_parser() -> argparse.ArgumentParser:
         "-O",
         "--outfile",
         type=str,
-        default="PLOM-scrap.pdf",
-        help="filename for output (optional, default PLOM-scrap.pdf)",
+        default="Plom-scrap.pdf",
+        help="filename for output (optional, default Plom-scrap.pdf)",
     )
     parser.add_argument(
         "-d",